requirements.txt
python-telegram-bot==20.7
python-dotenv
gspread
oauth2client
apscheduler
